pyproject.toml
# SPDX-FileCopyrightText: 2025 Max Mehl <https://mehl.mx>
#
# SPDX-License-Identifier: GPL-3.0-only

[tool.poetry]
name = "home-stream"
version = "0.4.1"
description = "Browser and streaming interface for local media files"
repository = "https://github.com/mxmehl/home-stream"
authors = ["Max Mehl <mail@mehl.mx>"]
readme = "README.md"
keywords = []
license = "GPL-3.0-only"
classifiers = []

[tool.poetry.scripts]
home-stream = "home_stream.app:main"

[tool.poetry.dependencies]
python = "^3.10"
bcrypt = "^4.3.0"
pyyaml = "^6.0.2"
flask = "^3.1.0"
flask-wtf = "^1.2.2"
flask-limiter = "^3.12"
redis = "^5.2.1"

[tool.poetry.group.web.dependencies]
uwsgi = "^2.0.28"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
pylint = "^3.3.6"
black = "^25.1.0"
mypy = "^1.15.0"
bump-my-version = "^1.1.1"
pytest = "^8.3.5"
pytest-flask = "^1.3.0"
pytest-cov = "^6.1.1"
beautifulsoup4 = "^4.13.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# FORMATTING settings
[tool.isort]
profile = "black"

[tool.black]
line-length = 100

# MYPY settings
[tool.mypy]
files = ["home_stream/*.py"]
disable_error_code = ["import-untyped"]

# Bump-My-Version
[tool.bumpversion]
commit = true
tag = true
allow_dirty = false
tag_name = "v{new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
regex = true
search = "^version = \"{current_version}\""
replace = "version = \"{new_version}\""
